feat/test: Add tokenizer to extract ingredientinfo
On a lot of recipe sites, there is additional information in the
ingredient listing like how the ingredient is supposed to be cut and the
quantity of the ingredient.
Add a tokenizer breaks down the listed ingredient info into its
components: Ingredient, state of ingredient, quantity, unit, brackets,
punctuation, etc.
Add a few preliminary tests.

The tokenizer will work with an parser and interpreter (yet to be
implemented) to create an Ingredient node and a Requirement edge.