fix/test: Change tokenizer operand-bracket-handling
There is a small error in reasoning when handling operands and brackets.
There should be an operand before the brackets and after (iff something
follows after the brackets) otherwise there won't be an operand between
the expression outside and inside of the brackets for the parser.

Change tokenizer test to fit the new bracket handling.