refactor: change preposition handling in tokenizer
Sometimes a state can also contain an amount and a unit. These states
usually have a preposition, e.g.: "200g chicken cut into 0.5 inch
strips". With the current preposition handling in the tokenizer 0.5 inch
would be information belonging to the ingredient, not the state.
So, for states followed by an preposition, the following tokens should
be merge into the state until another state or the end of the patterns
is reached.