fix: fix disappearing measurement records
If two measurements follow directly after each other and they are not
the measurement before the ingredient, one of the two disappear. The two
are merged together, so one of the two is in the altenative measurement
list. When they are merged with the ingredient and main measurement the
one in the list disappear since the list is not added to the main list.