refactor: refactor controllers for api testing
The MainController is used for debugging until the frontend and the REST
API is implemented. Rename MainController to DebugController. Add
Recipecontroller for API testing.

WARNING: The SecurityConfig was changed such that authentication is not
necessary for accessing the API. This needs to be removed once
connection to frontend is implemented.