Add react-frontend to the main branch (#9)
* feat: add login/register template and axios config

Create vite-react-project for frontend.
Add login and registration forms that are currently not submitting POSTs.
Add axios config to connect react frontend with spring-backend.

* feat: connect login, logout &  register to spring-backend

Add logout button. Add redirect to login after logout.
Connect login and registration submission to spring-backend. Add
redirect to homepage after login/registration.

* ui: add dynamic nav bar change depending on login status

Add a dynamic change in the navigation bar that switches the register
and login option with a profile and logout option depending on whether
the user is logged in or not.

* ui: Add upload page where users fill in recipe infos

Add page where users can fill out the name, description, ingredients and
instructions of a recipe they want to upload.
Add dynamic ingredient list and instruction list for easy editing and
removal of ingredients and instructions.

* ui: add cancel edit button for upload page

There is no way to cancel editing an item.

* refactor: Implement uploading via POST to Upload

The upload page's submit via POST request had not been implemented yet.

* refactor: remove username in upload

The username will be added in the backend, so there is no need for the
username to be in the post request.

* ui: add single recipe display to redirect to after upload

After the upload nothing indicates that the upload was successfull.
Add page that displays a recipe by its id that the upload can redirect
to. Redirect upload to recipe display page.

* chore: add TODO comments for error handling

Mark unimplemented error handling in Login and Register with TODO
comments.

* ui: Add page listing all recipes

* ui: Add page for scraping recipe website URLs

Add page for scraping a specific recipe from a recipe website via
submitting the link. Redirect scrape page to recipe display after
successful scraping.

* ui: add navbar buttons for scrape and all recipes

Add navbar buttons to access the scrape recipe and show all recipes
page.

* refactor: move axios_config.tsx to new config dir

* refactor: Move reusable types into types/ dir

* ui: add page that filters recipes by ingredients

Add page that allows user receive recipes filtered by the ingredients
the user wishes to use up.

* ui: fix navbar button text from scrape to filter

The filter button in the navbar has the text scrape.

* chore: update package-lock.json to fix vulnerabilities

* build: add react-frontend to docker compose

Add Dockerfile and .dockerignore for the react-frontend.
Update vite.config.ts for docker to work.
Add react-frontend to the docker-compose.yaml.