Update spring-backend for react-frontend (#8)
* refactor: Adapt RecipeController for frontend

The recipe controller's upload endpoint does not add the username to the
uploaded recipe that will be saved. The username is not contained in the
request from the body either.
There is no way to retrieve a single recipe by its id.
Add username extraction from jwtToken to RecipeController's
uploadRecipe.
Add endpoint where the frontend can get the a single recipe by its id.
Change the id generation of the models Recipe, Requirement, and AppUser
from internal Neo4j id to UUIDStringGenerator.class because the Neo4j
internal id contains double colons.

* refactor: wrap recipe URL to scrape for scrapeRecipe

Using just a string for a request body does not work.
Wrap the URL for the recipe site to scrape to fix this issue.

* refactor: Adapt to frontend port change to 3000

Change cross origin port for react frontend to 3000.
Move wrappers into separate wrappers/ directory.

* chore: move .gitattributes to project root dir

* refactor: update CI to test spring-backend on push